move history restoration to class
